folium_gui.py -> gui.py
